refactor: decompose page.go into gstate.go, content.go, walk.go, plaintext.go
Split the 725-line god file into focused units:
- gstate.go   — matrix, ident, gstate struct, xobjMaxDepth const
- content.go  — contentState interpreter + Content()
- walk.go     — walkState interpreter + walkTextBlocks()
- plaintext.go — plainTextState interpreter + GetPlainText(Page)

page.go retains Page struct, Reader nav API, page property accessors,
and the public GetPlainText/GetStyledTexts context-cancellation wrappers.
No behaviour changes; go build and go test -race pass.
