+++ gstate.go
@@ -1,0 +1,40 @@
+// Copyright 2014 The Go Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package pdf
+
+type matrix [3][3]float64
+
+var ident = matrix{{1, 0, 0}, {0, 1, 0}, {0, 0, 1}}
+
+func (x matrix) mul(y matrix) matrix {
+	var z matrix
+	for i := 0; i < 3; i++ {
+		for j := 0; j < 3; j++ {
+			for k := 0; k < 3; k++ {
+				z[i][j] += x[i][k] * y[k][j]
+			}
+		}
+	}
+	return z
+}
+
+type gstate struct {
+	Tc    float64
+	Tw    float64
+	Th    float64
+	Tl    float64
+	Tf    Font
+	Tfs   float64
+	Tmode int
+	Trise float64
+	Tm    matrix
+	Tlm   matrix
+	Trm   matrix
+	CTM   matrix
+}
+
+// xobjMaxDepth caps Form XObject recursion to guard against malformed PDFs
+// that contain cyclic or deeply nested XObject references.
+const xobjMaxDepth = 10
